mc33810: make mc33810getLiveData() constexpr
So it can be used in log_fields_generated.h
