+++ firmware/hw_layer/drivers/gpio/mc33810.cpp
@@ -23,8 +23,6 @@
 #include "gpio/gpio_ext.h"
 #include "gpio/mc33810.h"
 
-#include "mc33810_state_generated.h"
-
 #if EFI_PROD_CODE && (BOARD_MC33810_COUNT > 0)
 
 // For exti irq
@@ -40,19 +38,6 @@
 /*==========================================================================*/
 
 #define DRIVER_NAME				"mc33810"
-
-typedef enum {
-	MC33810_DISABLED = 0,
-	MC33810_WAIT_INIT,
-	MC33810_READY,
-	MC33810_FAILED
-} mc33810_drv_state;
-
-typedef enum {
-	COIL_IDLE = 0,
-	COIL_WAIT_SPARK_START,
-	COIL_WAIT_SPARK_END
-} mc33810_coil_state;
 
 #define MC_CMD_READ_REG(reg)			(0x0a00 | (((reg) & 0x0f) << 4))
 #define MC_CMD_SPI_CHECK				(0x0f00)
@@ -130,77 +115,7 @@ static THD_WORKING_AREA(mc33810_thread_wa, 256);
 #define INJ_MASK		0x0f
 #define IGN_MASK		0xf0
 
-/* Driver */
-struct Mc33810 : public GpioChip, public mc33810_state_s {
-	int init() override;
-
-	int writePad(size_t pin, int value) override;
-	brain_pin_diag_e getDiag(size_t pin) override;
-	void debug() override;
-
-	// internal functions
-	int spi_unselect();
-	int spi_rw(uint16_t tx, uint16_t* rx);
-	int spi_rw_array(const uint16_t *tx, uint16_t *rx, int n);
-	int update_output_and_diag();
-
-	int chip_init();
-	void wake_driver();
-
-	void ign_event(size_t pin, int value);
-	void on_spkdur(efitick_t now);
-
-	int chip_init_data();
-
-	const mc33810_config	*cfg;
-
-	/* cached output state - state last send to chip */
-	uint8_t					o_state_cached;
-	/* state to be sent to chip */
-	uint8_t					o_state;
-	/* direct driven output mask */
-	uint8_t					o_direct_mask;
-	/* IGN/GPGD mode bits: [7:4] - GP3..GP0 */
-	uint8_t 				o_gpgd_mask;
-
-	/* ALL STATUS RESPONSE value and flags */
-	bool					all_status_updated;
-	uint16_t				all_status_value;
-
-	/* OUTx fault registers */
-	uint16_t				out_fault[2];
-	/* GP mode fault register */
-	/* TODO: check documentation if these faults also applied to GPx outputs in IGN mode */
-	uint16_t				gp_fault;
-	/* IGN mode fault register */
-	uint16_t				ign_fault;
-
-	uint16_t				recentTx;
-
-	/* SPKDUR handling */
-	struct {
-		ioportid_t		port;
-		uint_fast8_t	pad;
-	} spkdur;
-	mc33810_coil_state 		coil_state;
-	uint8_t					active_coil_idx;	/* zero based, used as index of spark[] array */
-	uint8_t					spark_fault_mask;	/* 4 LSB bits are not used */
-	efitick_t				spartStart[MC33810_IGN_OUTPUTS];
-	int						spark_sync_err;
-
-	/* statistic */
-	int						rst_cnt;
-	int						cor_cnt;
-	int 					sor_cnt;
-	int 					ov_cnt;
-	int 					lv_cnt;
-
-	mc33810_drv_state		drv_state;
-
-	bool hadSuccessfulInit = false;
-};
-
-static Mc33810 chips[BOARD_MC33810_COUNT];
+Mc33810 mc33810_chips[BOARD_MC33810_COUNT];
 
 static const char* mc33810_pin_names[MC33810_OUTPUTS] = {
 	"mc33810.OUT1",		"mc33810.OUT2",		"mc33810.OUT3",		"mc33810.OUT4",
@@ -757,7 +672,7 @@ static THD_FUNCTION(mc33810_driver_thread, p) {
 		(void)msg;
 
 		for (int i = 0; i < BOARD_MC33810_COUNT; i++) {
-			auto chip = &chips[i];
+			auto chip = &mc33810_chips[i];
 
 			if (i == 0) {
 					engine->engineState.smartChipRestartCounter = chip->init_cnt;
@@ -957,7 +872,7 @@ int mc33810_add(brain_pin_e base, unsigned int index, const mc33810_config *cfg)
 	//if (cfg->spi_config.ssport == NULL)
 	//	return -1;
 
-	Mc33810& chip = chips[index];
+	Mc33810& chip = mc33810_chips[index];
 
 	/* already initted? */
 	if (chip.cfg != NULL)
@@ -1000,7 +915,7 @@ void mc33810_req_init() {
 	size_t i;
 
 	for (i = 0; i < BOARD_MC33810_COUNT; i++) {
-		auto& chip = chips[i];
+		auto& chip = mc33810_chips[i];
 
 		chip.need_init = true;
 	}
@@ -1024,12 +939,6 @@ case DWELL_8MS:
  return 0;
 }
 
-const mc33810_state_s* mc33810getLiveData(size_t idx) {
-	if (idx >= BOARD_MC33810_COUNT)
-		return nullptr;
-	return &chips[idx];
-}
-
 #else /* BOARD_MC33810_COUNT > 0 */
 
 int mc33810_add(brain_pin_e base, unsigned int index, const mc33810_config *cfg)
@@ -1037,10 +946,6 @@ int mc33810_add(brain_pin_e base, unsigned int index, const mc33810_config *cfg)
 	(void)base; (void)index; (void)cfg;
 
 	return -5;
-}
-
-const mc33810_state_s* mc33810getLiveData(size_t) {
-	return nullptr;
 }
 
 #endif /* BOARD_MC33810_COUNT */
